--- a/src/variationallaplace/optimization.jl
+++ b/src/variationallaplace/optimization.jl
@@ -373,35 +373,34 @@ function run_spDCM_iteration!(state::VLMTKState, setup::VLMTKSetup)
         end
     end
 
+    trmul(A, B) = sum(A' .* B)
     ϵ = reshape(y - f_μθ, ny)                   # error
     J = - dfdp   # Jacobian, unclear why we have a minus sign. Helmut: comes from deriving a Gaussian. 
     ## M-step: Fisher scoring scheme to find h = max{F(p,h)} // comment from MATLAB code
-    eltypeJ = ComplexF64
-    P = [spzeros(eltypeJ, size(Q[i])) for i = 1:nh]
-    PΣ = [spzeros(eltypeJ, size(Q[i])) for i = 1:nh]
-    JPJ = [spzeros(real(eltypeJ), size(J)) for i = 1:nh]
-    dFdλ = spzeros(eltypeJ, nh)
-    dFdλλ = spzeros(real(eltypeJ), nh, nh)
+    szQs = [size(Q[i]) for i ∈ 1:nh]
+    P = [spzeros(eltype(J), size(Q[i])) for i = 1:nh]
+    PΣ = [spzeros(eltype(J), size(Q[i])) for i = 1:nh]
+    JPJ = [spzeros(real(eltype(J)), size(J)) for i = 1:nh]
+    dFdλ = spzeros(eltype(J), nh)
+    dFdλλ = spzeros(real(eltype(J)), nh, nh)
     local iΣ, Σλ_po, Σθ_po, ϵ_λ
     for m = 1:8   # 8 seems arbitrary. Numbers of iterations taken from SPM code.
         iΣ = spzeros(eltype(J), ny, ny)
-        iΣ = sum(i ->Q[i]*exp(λ[i]), 1:nh)
-        # for i = 1:nh
-        #     iΣ .+= Q[i] * exp(λ[i])
-        # end
-        # iΣfac = qr(iΣ)
+        for i = 1:nh
+            iΣ .+= Q[i] * exp(λ[i])
+        end
         Pp = real(J' * iΣ * J)    # in the SPM code 'real()' is applied here and there. Essentially the complex dimension seems to be ignored at some point with no rational given. Is that okay?
         Σθ_po = inv(Pp + Πθ_pr)
         if nh > 1
             for i = 1:nh
                 P[i] = Q[i]*exp(λ[i])
-                PΣ[i] = iΣ \ P[i]              # why is PΣ still type stable even if iΣ is of type Any? Maybe still slowing down since it needs to infer type.
+                PΣ[i] = iΣ \ P[i]
                 JPJ[i] = real(J'*P[i]*J)
             end
             for i = 1:nh
-                dFdλ[i] = (tr(PΣ[i]) - real(dot(ϵ, P[i], ϵ)) - (Σθ_po' ⋅ JPJ[i]))/2
+                dFdλ[i] = (tr(PΣ[i]) - real(dot(ϵ, P[i], ϵ)) - trmul(Σθ_po,  JPJ[i]))/2
                 for j = i:nh
-                    dFdλλ[i, j] = -real(PΣ[i]' ⋅ PΣ[j])/2   #tr(PΣ[:,:,i] * PΣ[:,:,j])
+                    dFdλλ[i, j] = -real(trmul(PΣ[i], PΣ[j]))/2
                     dFdλλ[j, i] = dFdλλ[i, j]
                 end
             end