@MasonProtter 2025-05-16
Main things I did:

- Get rid of MTKParams usage -- the second argument to f_model is the tunable parameters, and the third argument is the non-tunable parameters, so I just passed those in directly
- Fixing an annoying type instability that was caused by writing `eltypeJ = ComplexFloat64`, and then using that value in closures. Went back to just calling `eltype(J)` which relieved some type instabilities
- Replaced `dot(A', B)` with `sum(A' .* B)` because unfortunately, the dot product with a adjoint of a sparse array hits a slow fallback. 
- Added a function barrier to control the type instabilities in `eigen`
- Fused some operations in `transfer_function` to save on RAM and not have to pass over the same data 3 times.