@MasonProtter 2025-05-16
I'm getting 5 regions in ~5.5 seconds.